post second product

+++ src/main/java/org/factoriaf5/store/model/Product.java
@@ -2,16 +2,16 @@ package org.factoriaf5.store.model;
 
 public class Product {
     private final String name;
-    private final int price;
+    private final double price;
 
-public Product(String name, int price) {
+public Product(String name, double price) {
     this.name = name;
     this.price = price;
 }
 public String getName() {
     return name;
 }
-public int getPrice() {
+public double getPrice() {
     return price;
 }
 }
